update the array size type from size_t to int avoiding bugs with negative args

+++ search.h
@@ -3,6 +3,6 @@
 
 int sequential_search(int *array, int k, size_t array_size);
 int sorted_sequential_search(int *array, int k);
-int binary_search(int *array, int k, size_t array_size);
+int binary_search(int *array, int k, int array_size);
 
 #endif
